# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppStoreCat API",
        "description": "Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:7460/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/apps/{platform}/{externalId}/keywords": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "Get keyword density for an app",
                "operationId": "appKeywords",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en-US"
                        }
                    },
                    {
                        "name": "ngram",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1,
                            "enum": [
                                1,
                                2,
                                3
                            ]
                        }
                    },
                    {
                        "name": "version_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Keyword density list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/KeywordDensityResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "App not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/apps/{platform}/{externalId}/keywords/compare": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "Compare keyword density with other apps",
                "operationId": "compareKeywords",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "externalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "app_ids",
                        "in": "query",
                        "description": "App IDs to compare (max 5)",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "version_ids",
                        "in": "query",
                        "description": "Version ID per app (keyed by app ID)",
                        "required": false,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ngram",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2,
                                3
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Keyword comparison data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KeywordCompareResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "App not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "KeywordCompareResource": {
                "required": [
                    "apps",
                    "keywords"
                ],
                "properties": {
                    "apps": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "icon_url": {
                                    "type": "string",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "keywords": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "KeywordDensityResource": {
                "required": [
                    "keyword",
                    "count",
                    "density"
                ],
                "properties": {
                    "language": {
                        "type": "string",
                        "example": "en-US"
                    },
                    "ngram_size": {
                        "type": "integer",
                        "example": 2
                    },
                    "keyword": {
                        "type": "string",
                        "example": "photo editor"
                    },
                    "count": {
                        "type": "integer",
                        "example": 5
                    },
                    "density": {
                        "type": "number",
                        "format": "float",
                        "example": 2.35
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentication using Bearer token. Example: Bearer {your-token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}